// src/types/threshold.ts
export type Threshold = {
    id: number;
    deviceId: number; 
    deviceName: string;
    thresholdVelocityX: number;
    messageThresholdVelocityX: string;
    thresholdVelocityY: number;
    messageThresholdVelocityY: string;
    thresholdVelocityZ: number;
    messageThresholdVelocityZ: string;
    thresholdAccelerationX: number;
    messageThresholdAccelerationX: string;
    thresholdAccelerationZ: number;
    messageThresholdAccelerationZ: string;
    thresholdTemperature: number;
    messageThresholdTemperature: string;
    createdAt: string;
}